pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "vtes-rulings"
dynamic = ["version"]
authors = [{ name = "VEKN" }]
description = "Website for VTES cards rulings"
keywords = ["vtes", "Vampire: The Eternal Struggle", "CCG"]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Development Status :: 0 - Production/Stable",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Framework :: Flask",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
]
dependencies = [
    "flask>=3",
    "flask-babel>=4",
    "unidecode>=1",
    "Jinja2>=3",
    "krcg",
]

[project.optional-dependencies]
dev = [
    "black>=24",
    "build>=1",
    "check-manifest>=0",
    "ipython>=8",
    "pytest>=8",
    "requests>=2",
    "ruff>=0",
    "setuptools_scm>=8",
    "twine>=5",
    "yamllint",
]

[project.scripts]
rulings-web = "vtesrulings:main"

[project.urls]
Repository = "https://github.com/vtes-biased/vtes-rulings"

[tool.setuptools.packages.find]
where = ["src"]
include = ["vtesrulings*"]

[tool.setuptools.package-data]
"*" = ["*.yaml"]

[tool.setuptools_scm]
